+++ fiddles/angular2/fiddle-0048-SlickGridTree/app/components/acme-grid/services/index.ts
@@ -1,0 +1,4 @@
+export * from './acme-grid-columns.service';
+export * from './acme-grid-event.service';
+export * from './acme-grid-resize.service';
+export * from './acme-grid-rows.service';