+++ fiddles/angular2/fiddle-0048-SlickGridTree/app/components/ms-slick-grid/index.ts
@@ -1,5 +1,0 @@
-export * from './gridsync.service';
-export * from './interfaces';
-export * from './selection-model';
-export * from './slick-grid';
-export * from './virtualized-collection';