+++ fiddles/angular2/fiddle-0048-SlickGridTree/app/fatal-error.ts
@@ -1,5 +1,5 @@
 var fatalError = {
-  write: function(err) {
+  write: function (err) {
     document.write('<H3>' + err + '</H3>');
   }
 };