+++ fiddles/angular2/fiddle-0048-SlickGridTree/app/header/header.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'header',
-    moduleId: module.id,
-    templateUrl: './app/header/header.component.html',
-    styleUrls: ['./app/header/header.component.css']
-})
-export class HeaderComponent { }