+++ fiddles/angular2/fiddle-0048-SlickGridTree/app/main.ts
@@ -1,5 +1,5 @@
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-import { AppModule } from './app.module';
+import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
+import {AppModule} from './app.module';
 
 import * as meta from './meta';
 import * as fatalError  from './fatal-error';