+++ fiddles/angular2/fiddle-0048-SlickGridTree/app/meta.ts
@@ -1,7 +1,7 @@
 var metadata = {
-     fiddleHeader: 'Angular 2 - fiddle-0048-SlickGridTree',
-     urls: {
-          github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0048-SlickGridTree'
-     },
-     consoleTag: 'H O U S E  ~  f i d d l e s'
+  fiddleHeader: 'Angular 2 - fiddle-0048-SlickGridTree',
+  urls: {
+    github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0048-SlickGridTree'
+  },
+  consoleTag: 'H O U S E  ~  f i d d l e s'
 };