+++ fiddles/angular2/fiddle-0048-SlickGridTree/systemjs-local.config.js
@@ -18,6 +18,7 @@
     'compiler',
     'core',
     'http',
+    'forms',
     'platform-browser',
     'platform-browser-dynamic',
     'router',