+++ fiddles/angular2/fiddle-0050-CssElementQueries/app/app.component.ts
@@ -1,0 +1,20 @@
+import {
+    Component,
+} from '@angular/core';
+
+
+@Component({
+  selector: 'app',
+  moduleId: module.id,
+  templateUrl: './app/app.component.html',
+  styleUrls: ['./app/app.component.css']
+})
+export class AppComponent {
+  options:any;
+
+  constructor() {
+    this.options = {
+      src: 'app/surface.jpg'
+    };
+  }
+}