+++ fiddles/angular2/fiddle-0050-CssElementQueries/app/app.module.ts
@@ -1,0 +1,11 @@
+import { NgModule }      from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { FillImageModule } from './components/fill-image/fill-image.module';
+import { AppComponent }  from './app.component';
+
+@NgModule({
+  imports:      [ BrowserModule, FillImageModule],
+  declarations: [ AppComponent ],
+  bootstrap:    [ AppComponent ]
+})
+export class AppModule { }