+++ fiddles/angular2/fiddle-0050-CssElementQueries/app/components/fill-image/fill-image.module.ts
@@ -1,0 +1,19 @@
+
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FillImageComponent } from './fill-image.component';
+
+@NgModule({
+    imports: [
+        CommonModule
+    ],
+    declarations: [
+        FillImageComponent
+    ],
+    exports: [
+        FillImageComponent
+    ]
+})
+export class FillImageModule {
+
+}