+++ fiddles/angular2/template/systemjs.config.js
@@ -1,60 +1,36 @@
-/**
- * PLUNKER VERSION (based on systemjs.config.js in angular.io)
- * System configuration for Angular 2 samples
- * Adjust as necessary for your application needs.
- */
-(function(global) {
+(function (global) {
 
-  var ngVer = '@4.0.2'; // lock in the angular package version; do not let it float to current!
+  var ngVer = '@4.0.2',
+    map = {
+      'app': 'app',
+      '@angular': 'https://npmcdn.com/@angular',
+      'angular-in-memory-web-api': 'https://npmcdn.com/angular-in-memory-web-api@0.3.1/bundles/in-memory-web-api.umd.js',
+      'rxjs': 'https://npmcdn.com/rxjs@5.3.0',
+      'ts': 'https://npmcdn.com/plugin-typescript@7.0.6/lib/plugin.js',
+      'typescript': 'https://npmcdn.com/typescript@2.2.2/lib/typescript.js',
+    },
+    packages = {
+      'app': {main: 'main.ts', defaultExtension: 'ts'},
+      'rxjs': {defaultExtension: 'js'},
+      'angular-in-memory-web-api': {defaultExtension: 'js'}
+    },
+    ngPackageNames = [
+      'common',
+      'compiler',
+      'core',
+      'http',
+      'platform-browser',
+      'platform-browser-dynamic',
+      'router',
+      'router-deprecated',
+      'upgrade',
+    ];
 
-  //map tells the System loader where to look for things
-  var  map = {
-    'app':                        'app',
-
-    '@angular':                   'https://npmcdn.com/@angular', // sufficient if we didn't pin the version
-    'angular2-in-memory-web-api': 'https://npmcdn.com/angular2-in-memory-web-api', // get latest
-    'rxjs':                       'https://npmcdn.com/rxjs@5.0.0-beta.6',
-    'ts':                         'https://npmcdn.com/plugin-typescript@4.0.10/lib/plugin.js',
-    'typescript':                 'https://npmcdn.com/typescript@1.8.10/lib/typescript.js',
-  };
-
-  //packages tells the System loader how to load when no filename and/or no extension
-  var packages = {
-    'app':                        { main: 'main.ts',  defaultExtension: 'ts' },
-    'rxjs':                       { defaultExtension: 'js' },
-    'angular2-in-memory-web-api': { defaultExtension: 'js' },
-  };
-
-  var ngPackageNames = [
-    'common',
-    'compiler',
-    'core',
-    'http',
-    'platform-browser',
-    'platform-browser-dynamic',
-    'router',
-    'router-deprecated',
-    'upgrade',
-  ];
-
-  // Add map entries for each angular package
-  // only because we're pinning the version with `ngVer`.
-  ngPackageNames.forEach(function(pkgName) {
-    map['@angular/'+pkgName] = 'https://npmcdn.com/@angular/' + pkgName + '@' + ngVer + '/bundles/' + pkgName + '.umd.js';
-  });
-
-  // Add package entries for angular packages
-  ngPackageNames.forEach(function(pkgName) {
-
-    // Bundled (~40 requests):
-    packages['@angular/'+pkgName] = { main: pkgName + '.umd.js', defaultExtension: 'js' };
-
-    // Individual files (~300 requests):
-    //packages['@angular/'+pkgName] = { main: 'index.js', defaultExtension: 'js' };
+  ngPackageNames.forEach(function (pkgName) {
+    map['@angular/' + pkgName] = 'https://npmcdn.com/@angular/' + pkgName + ngVer + '/bundles/' + pkgName + '.umd.js';
   });
 
   var config = {
-    // DEMO ONLY! REAL CODE SHOULD NOT TRANSPILE IN THE BROWSER
     transpiler: 'typescript',
     typescriptOptions: {
       emitDecoratorMetadata: true